dirty unfinished export endpoint; ++ endpoint for available classes; ++ dirty endpoint to register dated class; ++ minor fixes/renamings

+++ appapi/src/main/java/api/common/SimpleCUDResource.java
@@ -1,16 +1,13 @@
 package api.common;
 
 import com.google.gson.JsonSyntaxException;
-import com.google.gson.reflect.TypeToken;
 import core.GenericCUDService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.ws.rs.core.Response;
-import java.lang.reflect.Type;
 import java.math.BigInteger;
 import java.sql.SQLException;
-import java.util.Map;
 
 public abstract class SimpleCUDResource<T> extends CommonResource {
     private final String SCHEMA_NAME = "fcimapp";
@@ -33,8 +30,14 @@ public abstract class SimpleCUDResource<T> extends CommonResource {
     }
 
     public Response add(String json){
+        log.info("add request received: <>", json);
+
         if (!cudSeviceInitialized) {
             return Response.status(500).build();
+        }
+        T obj = gson.fromJson(json, clazz);
+        if (obj == null) {
+            return Response.status(400).entity("Invalid input data").build();
         }
         try {
             json = (json == null ? "" : json);
@@ -48,12 +51,21 @@ public abstract class SimpleCUDResource<T> extends CommonResource {
         }
     }
 
-    public Response update(String params){
-        Type paramsType = new TypeToken<Map<String, String>>() {}.getType();
-        Map<String, String> parsedParams = gson.fromJson(params, paramsType);
-        Long id = Long.valueOf(parsedParams.get("pk"));
-        String fieldName = parsedParams.get("name");
-        String fieldValue = parsedParams.get("value");
+    public Response update(Long id, String fieldName, String fieldValue){
+        log.info("update received: <{}, {}, {}>", id, fieldName, fieldValue);
+
+//        log.info("update received: <{}>", params);
+//        Type paramsType = new TypeToken<Map<String, String>>() {}.getType();
+//        Map<String, String> parsedParams = gson.fromJson(params, paramsType);
+//        if (parsedParams == null
+//                || !parsedParams.containsKey("pk")
+//                || !parsedParams.containsKey("value")
+//                || !parsedParams.containsKey("name")) {
+//            return Response.status(400).entity("Invalid input data").build();
+//        }
+//        Long id = Long.valueOf(parsedParams.get("pk"));
+//        String fieldName = parsedParams.get("name");
+//        String fieldValue = parsedParams.get("value");
 
         if (!cudSeviceInitialized) {
             return Response.status(500).build();
@@ -70,6 +82,9 @@ public abstract class SimpleCUDResource<T> extends CommonResource {
     public Response delete(Long id){
         if (!cudSeviceInitialized) {
             return Response.status(500).build();
+        }
+        if (id == null || id <= 0) {
+            return Response.status(400).entity("Invalid input data").build();
         }
         try {
             cudService.delete(id);
